# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: NFL Picks League API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /leagues/{leagueId}/events:
    get:
      operationId: listPickEvents
      tags: [pick-events]
      summary: List pick events for a league
      parameters:
        - $ref: "#/components/parameters/LeagueId"
      responses:
        "200":
          description: Events
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PickEvent"
        "403":
          $ref: "#/components/responses/Forbidden"
    post:
      operationId: createPickEvent
      tags: [pick-events]
      summary: Create a pick event (commissioner only)
      parameters:
        - $ref: "#/components/parameters/LeagueId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PickEventInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PickEvent"
        "400":
          $ref: "#/components/responses/BadRequest"
        "403":
          $ref: "#/components/responses/Forbidden"
  /leagues/{leagueId}/events/{eventId}:
    get:
      operationId: getPickEvent
      tags: [pick-events]
      summary: Get pick event detail
      parameters:
        - $ref: "#/components/parameters/LeagueId"
        - $ref: "#/components/parameters/EventId"
      responses:
        "200":
          description: Event
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PickEvent"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
    patch:
      operationId: updatePickEvent
      tags: [pick-events]
      summary: Update pick event (commissioner only)
      parameters:
        - $ref: "#/components/parameters/LeagueId"
        - $ref: "#/components/parameters/EventId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PickEventUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PickEvent"
        "400":
          $ref: "#/components/responses/BadRequest"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
  /leagues/{leagueId}/events/{eventId}/lock:
    post:
      operationId: lockPickEvent
      tags: [pick-events]
      summary: Lock lines and publish event (commissioner only)
      parameters:
        - $ref: "#/components/parameters/LeagueId"
        - $ref: "#/components/parameters/EventId"
      responses:
        "200":
          description: Event locked and published
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PickEvent"
        "400":
          $ref: "#/components/responses/BadRequest"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
  /leagues/{leagueId}/events/{eventId}/finalize:
    post:
      operationId: finalizePickEvent
      tags: [pick-events]
      summary: Finalize event results (commissioner only)
      parameters:
        - $ref: "#/components/parameters/LeagueId"
        - $ref: "#/components/parameters/EventId"
      responses:
        "200":
          description: Event finalized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PickEvent"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"

  # ── Event Games ──────────────────────────────────────────────────────────────
...
  /leagues/{leagueId}/standings:
    get:
      operationId: getSeasonStandings
      tags: [standings]
      summary: Get season standings for a league
      parameters:
        - $ref: "#/components/parameters/LeagueId"
      responses:
        "200":
          description: Season standings
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/StandingEntry"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
  /leagues/{leagueId}/events/{eventId}/standings:
    get:
      operationId: getEventStandings
      tags: [standings]
      summary: Get standings for a single pick event
      parameters:
        - $ref: "#/components/parameters/LeagueId"
        - $ref: "#/components/parameters/EventId"
      responses:
        "200":
          description: Event standings
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/EventStandingEntry"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"

  # ── Dashboard ────────────────────────────────────────────────────────────────
...
components:
  parameters:
...
    LeagueId:
      name: leagueId
      in: path
      required: true
      schema:
        type: number
    EventId:
      name: eventId
      in: path
      required: true
      schema:
        type: number
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ApiError"
...
    Forbidden:
      description: Forbidden
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ApiError"
    NotFound:
      description: Not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ApiError"
  schemas:
...
    ApiError:
      type: object
      required: [error]
      properties:
        error:
          type: string

    # ── Auth ──────────────────────────────────────────────────────────────────
...
    PickEvent:
      type: object
      required: [id, leagueId, name, nflWeek, nflSeason, status, submissionDeadline, revealAt, createdAt]
      properties:
        id:
          type: number
        leagueId:
          type: number
        name:
          type: string
        nflWeek:
          type: number
        nflSeason:
          type: number
        status:
          type: string
          enum: [draft, open, locked, revealed, finalized]
        submissionDeadline:
          type: string
          format: date-time
        revealAt:
          type: string
          format: date-time
        tiebreakerQuestion:
          type: ["string", "null"]
        tiebreakerResult:
          type: ["number", "null"]
        notes:
          type: ["string", "null"]
        submissionCount:
          type: number
        totalMembers:
          type: number
        publishedAt:
          type: ["string", "null"]
          format: date-time
        finalizedAt:
          type: ["string", "null"]
          format: date-time
        createdAt:
          type: string
          format: date-time
    PickEventInput:
      type: object
      required: [name, nflWeek, nflSeason, submissionDeadline, revealAt]
      properties:
        name:
          type: string
          minLength: 1
        nflWeek:
          type: number
        nflSeason:
          type: number
        submissionDeadline:
          type: string
          format: date-time
        revealAt:
          type: string
          format: date-time
        tiebreakerQuestion:
          type: string
        notes:
          type: string
    PickEventUpdate:
      type: object
      properties:
        name:
          type: string
          minLength: 1
        submissionDeadline:
          type: string
          format: date-time
        revealAt:
          type: string
          format: date-time
        tiebreakerQuestion:
          type: ["string", "null"]
        tiebreakerResult:
          type: ["number", "null"]
        notes:
          type: ["string", "null"]

    # ── Event Games ───────────────────────────────────────────────────────────
...
    StandingEntry:
      type: object
      required: [userId, displayName, profileImageUrl, totalPoints, eventsEntered, weeklyWins, normalCorrect, normalTotal, moneyCorrect, moneyTotal, rank]
      properties:
        userId:
          type: string
        displayName:
          type: ["string", "null"]
        profileImageUrl:
          type: ["string", "null"]
        totalPoints:
          type: number
        eventsEntered:
          type: number
        weeklyWins:
          type: number
        normalCorrect:
          type: number
        normalTotal:
          type: number
        moneyCorrect:
          type: number
        moneyTotal:
          type: number
        rank:
          type: number
    EventStandingEntry:
      type: object
      required: [userId, displayName, profileImageUrl, points, maxPossible, normalCorrect, normalTotal, moneyCorrect, moneyTotal, rank, isEliminated, tiebreakerAnswer]
      properties:
        userId:
          type: string
        displayName:
          type: ["string", "null"]
        profileImageUrl:
          type: ["string", "null"]
        points:
          type: number
        maxPossible:
          type: number
        normalCorrect:
          type: number
        normalTotal:
          type: number
        moneyCorrect:
          type: number
        moneyTotal:
          type: number
        rank:
          type: number
        isEliminated:
          type: boolean
        tiebreakerAnswer:
          type: ["number", "null"]

    # ── Dashboard ─────────────────────────────────────────────────────────────
...